app/test_requirements.txt
pytest-cov
pytest-django
pytest-factoryboy
allure-pytest
pytest-xdist
